FIX: made ad-hoc changes for Tcl 9.0 compatibility

+++ generic/tclXprocess.c
@@ -82,7 +82,7 @@ TclX_ExeclObjCmd (ClientData clientData,
     char **argList = staticArgv;
     int nextArg = 1;
     char *argStr;
-    int argObjc;
+    long int argObjc;
     Tcl_Obj **argObjv;
     char *path, *argv0 = NULL;
     int idx, status;
